requirements.txt
django-mysql
mysqlclient
djangorestframework
